Change from named volume to mounted folder from host for simplicity and more sane file linking.

+++ script/start-server.sh
@@ -7,18 +7,20 @@ set -euo pipefail
 script_dir="$(cd "$(dirname "$0")" && env pwd --physical)"
 source_dir="$(readlink --canonicalize "$script_dir/..")"
 docker_dir="$source_dir/docker"
+server_dir="$source_dir/server-files"
 
 help() {
 	cat <<-EOF
 		Usage: $(basename "$0") [ -u ]
 		  -h, --help    Prints this message.
 		  -u, --update  Updates the server files before starting the server.
+		  -p, --update-only  Updates the server files and exits. Implies --update.
 	EOF
 }
 
 canonicalized=$(getopt --name "$(basename "$0")" \
-	--options hu \
-	--longoptions help,update \
+	--options hup \
+	--longoptions help,update,update-only \
 	-- "$@") || status=$?
 
 if [ "${status-0}" -ne 0 ]; then
@@ -37,15 +39,20 @@ for arg in "$@"; do
 	-u | --update)
 		update_server=true
 		;;
+	-p | --update-only)
+		update_server=true
+		update_only=true
+		;;
 	esac
 done
 
 update_server=${update_server-false}
+update_only=${update_only-false}
 
 if $update_server; then
 	if ! sudo --non-interactive true 2>/dev/null; then
 		echo sudo password required to update server files.
-		echo This is to fix server file permissions after updating.
+		echo Required to set server file permissions.
 		sudo --validate || exit
 	fi
 fi
@@ -68,27 +75,21 @@ if $update_server; then
 
 	"$script_dir/fix-permissions.sh"
 
-	mount_dir="$("$script_dir/cd-mountpoint.sh" --path)"
-
 	echo Linking host files to volume...
 
-	mkdir --parents "$mount_dir/server"
-	ln --logical --force "$source_dir/cfg/start.sh" "$mount_dir/server/start.sh"
-
-	echo Linking volume files to host...
-
-	ln --symbolic --no-dereference --force "$mount_dir" "$source_dir/server-files"
-
-	# handle=docker container ls -all --quiet --filter name=server-files
+	mkdir --parents "$server_dir/server"
+	ln --logical --force "$source_dir/cfg/start.sh" "$server_dir/server/start.sh"
 fi
 
-compose_run=("${compose[@]}" --progress plain run --rm --service-ports palworld-server)
+if ! $update_only; then
+	compose_run=("${compose[@]}" --progress plain run --rm --service-ports palworld-server)
 
-log="$logs_dir/log-$(date +%Y%j-%H%M%S).txt"
-mkdir --parents "$(dirname "$log")"
+	log="$logs_dir/log-$(date +%Y%j-%H%M%S).txt"
+	mkdir --parents "$(dirname "$log")"
 
-echo Running command: "${compose_run[*]}"
-echo "... with output logging to file: $log"
-echo "... in screen daemon; screen -r palworld"
+	echo Running command: "${compose_run[*]}"
+	echo "... with output logging to file: $log"
+	echo "... in screen daemon; screen -r palworld"
 
-screen -dmS palworld -L -Logfile "$log" "${compose_run[@]}"
+	screen -dmS palworld -L -Logfile "$log" "${compose_run[@]}"
+fi
